requirements.txt
streamlit==1.9.0
diffractsim==0.2.4
matplotlib==3.5.1
numpy==1.22.3
Pillow==9.0.1
opencv-python-headless==4.5.5.64
